requirements.txt
asgiref==3.8.1
Django==5.0.4
django-simple-history==3.5.0
pipfreeze==2.0.3
sqlparse==0.5.0
typing_extensions==4.11.0
pillowasgiref==3.8.1
Django==5.0.4
django-simple-history==3.5.0
djangorestframework==3.15.2
pillow==10.3.0
pipfreeze==2.0.3
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
